Basic failing state machine.

+++ app/controllers/user_sessions_controller.rb
@@ -2,7 +2,6 @@ class UserSessionsController < ApplicationController
   respond_to :html
 
   before_filter :require_no_user, only: %w(new create)
-  before_filter :require_user, only: "destroy"
 
   def new
     respond_with(@user_session = UserSession.new)
@@ -11,15 +10,15 @@ class UserSessionsController < ApplicationController
   def create
     @user_session = UserSession.new params[:user_session]
     if @user_session.save
-      redirect_back_or_default edit_user_path, notice: "Login successful!"
+      redirect_back_or_default profile_path(@user_session.user),
+        notice: "Login successful!"
     else
       render "pages/root"
     end
   end
 
   def destroy
-    current_user_session.destroy
-    redirect_back_or_default new_user_session_path,
-      notice: "Logout successful!"
+    current_user_session.destroy if current_user_session
+    redirect_back_or_default root_path
   end
 end
